+++ lopatina.sofia/P3/add_numbers.hpp
@@ -1,0 +1,10 @@
+#ifndef ADD_NUMBERS_HPP
+#define ADD_NUMBERS_HPP
+
+#include <iostream>
+namespace lopatina
+{
+  char * addNumbers(char * first_str, const char * second_str, size_t size1, size_t size2);
+}
+
+#endif
