+++ lopatina.sofia/P3/define_len.cpp
@@ -1,0 +1,10 @@
+#include "define_len.hpp"
+
+size_t lopatina::defineLen(const char * str)
+{
+  size_t count = 0;
+  while (str[count++] != 0)
+  {
+  }
+  return count;
+}
