+++ lopatina.sofia/P3/define_len.hpp
@@ -1,0 +1,10 @@
+#ifndef DEFINE_LEN_HPP
+#define DEFINE_LEN_HPP
+
+#include <iostream>
+namespace lopatina
+{
+  size_t defineLen(const char * str);
+}
+
+#endif
