+++ lopatina.sofia/P3/exclude_common.hpp
@@ -1,0 +1,10 @@
+#ifndef EXCLUDE_COMMON_HPP
+#define EXCLUDE_COMMON_HPP
+
+#include <iostream>
+namespace lopatina
+{
+  char * excludeCommon(char * first_str, const char * second_str, size_t size1, const size_t size2);
+}
+
+#endif
