+++ lopatina.sofia/P3/expand_array.hpp
@@ -1,0 +1,10 @@
+#ifndef EXPAND_ARRAY_HPP
+#define EXPAND_ARRAY_HPP
+
+#include <iostream>
+namespace lopatina
+{
+  char * expandArray(char * arr, size_t size);
+}
+
+#endif
